optimization: Agent 安装依赖接口增加重试机制 (closed #2012)
