+++ apps/backend/components/collections/agent_new/base.py
@@ -11,12 +11,10 @@ specific language governing permissions and limitations under the License.
 import abc
 import random
 import re
-import time
 from collections import defaultdict
 from dataclasses import dataclass
-from typing import Any, Callable, Dict, List, Match, Optional, Set, Tuple, Type, Union
+from typing import Any, Dict, List, Match, Optional, Set, Tuple, Union
 
-import wrapt
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
@@ -332,54 +330,6 @@ class AgentBaseService(BaseService, metaclass=abc.ABCMeta):
         else:
             host_ap: models.AccessPoint = common_data.host_id__ap_map[host.bk_host_id]
         return host_ap
-
-
-class RetryHandler:
-    """重试处理器"""
-
-    # 重试间隔
-    interval: float = None
-    # 重试次数
-    retry_times: int = None
-    # 需要重试的异常类型
-    exception_types: List[Type[Exception]] = None
-
-    def __init__(
-        self, interval: float = 0, retry_times: int = 1, exception_types: Optional[List[Type[Exception]]] = None
-    ):
-        self.interval = max(interval, 0)
-        self.retry_times = max(retry_times, 0)
-        self.exception_types = exception_types or [Exception]
-
-    @wrapt.decorator
-    def __call__(self, wrapped: Callable, instance: Optional[object], args: Tuple[Any], kwargs: Dict[str, Any]):
-        """
-        :param wrapped: 被装饰的函数或类方法
-        :param instance:
-            - 如果被装饰者为普通类方法，该值为类实例
-            - 如果被装饰者为 classmethod / 类方法，该值为类
-            - 如果被装饰者为类/函数/静态方法，该值为 None
-        :param args: 位置参数
-        :param kwargs: 关键字参数
-        :return:
-        """
-        call_times = self.retry_times + 1
-        while call_times > 0:
-            call_times = call_times - 1
-            try:
-                return wrapped(*args, **kwargs)
-            except Exception as exc_val:
-                # 重试次数已用完或者异常捕获未命中时，抛出原异常
-                if call_times == 0 or not self.hit_exceptions(exc_val):
-                    raise
-                # 休眠一段时间
-                time.sleep(self.interval)
-
-    def hit_exceptions(self, exc_val: Exception) -> bool:
-        for exception in self.exception_types:
-            if isinstance(exc_val, exception):
-                return True
-        return False
 
 
 # 根据 JOB 的插件额外封装一层，保证后续基于 Agent 增加定制化功能的可扩展性
